code cleanup: remove unneccessary semicolons, casts etc.
git-svn-id: https://svn.codehaus.org/mule/branches/mule-3.x@24840 bf997673-6b11-0410-b953-e057580c5b09

+++ core/src/main/java/org/mule/util/store/PartitionedInMemoryObjectStore.java
@@ -174,7 +174,7 @@ public class PartitionedInMemoryObjectStore<T extends Serializable> extends Abst
 
         while ((oldestEntry = store.firstEntry()) != null)
         {
-            Long oldestKey = (Long) oldestEntry.getKey();
+            Long oldestKey = oldestEntry.getKey();
             long oldestKeyValue = oldestKey.longValue();
 
             if (TimeUnit.NANOSECONDS.toMillis(now - oldestKeyValue) >= entryTTL)
